Move files, link up ingredients page, and add some styling
