+++ client/src/App.js
@@ -1,4 +1,5 @@
-import Register from "./components/Register";
+import Register from "./components/pages/Register";
+import Ingredients from "./components/pages/Ingredients"
 import "./App.css";
 import { BrowserRouter as Router, Switch, Route } from "react-router-dom";
 import Navigation from "./components/Navigation";
@@ -11,6 +12,9 @@ function App() {
                 <Switch>
                     <Route exact path="/register">
                         <Register />
+                    </Route>
+                    <Route exact path="/ingredients">
+                        <Ingredients />
                     </Route>
                 </Switch>
             </Router>
