+++ client/src/components/pages/Ingredients.js
@@ -1,20 +1,21 @@
 import React from "react";
 // import Navbar from "../components/Navbar.js"
 import Button from "react-bootstrap/Button";
-import IngredientsList from "../components/IngredientsList";
+import IngredientsList from "../IngredientsList";
 
 const Ingredients = () => {
     // TODO: Attach to a form component which pushes ingredient to the database on submit
     const addIngredient = () => {};
 
     return (
-        <div>
+        <div className="mx-auto w-75 text-center">
             {/* TODO: Add Navbar component here:
             <Navbar /> */}
-            <h1>Your Ingredients</h1>
-            <Button variant="primary" onClick={() => addIngredient()}>
+            <h1>Your Ingredients
+            <Button variant="primary" className="mx-5" onClick={() => addIngredient()}>
                 Add Ingredient
             </Button>
+            </h1>
             <IngredientsList />
         </div>
     );
